refactor: Update authRepository import to use old-auth-repository
This commit updates the import statement in the OboardingForm.tsx file to use the old-auth-repository instead of the auth-repository. This change is necessary due to a recent refactoring in the authentication feature.

Recent user commits:
- Add lazy loading for images in profile page
- Add LoadingPage component
- Add lucide-react dependency for using User icon in UserProfile component
- Update UserProfile component to save user data
- Add useSaveUser hook for saving user data
- Update ConvexDatabase saveUserMutation to include user ID and user object
- Update user query in saveUserMutation to use "_id" field instead of "userId"

Recent repository commits:
- Add lazy loading for images in profile page
- Add LoadingPage component
- Add lucide-react dependency for using User icon in UserProfile component
- Update UserProfile component to save user data
- Add useSaveUser hook for saving user data
- Update ConvexDatabase saveUserMutation to include user ID and user object
- Update user query in saveUserMutation to use "_id" field instead of "userId"
- Add saveUser method to ProfileUseCases
- Add saveUser method to ProfileRepository
